+++ sdk/log/logger.go
@@ -10,7 +10,12 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/log"
 	"go.opentelemetry.io/otel/log/embedded"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/sdk"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
+	"go.opentelemetry.io/otel/sdk/log/internal/x"
+	semconv "go.opentelemetry.io/otel/semconv/v1.36.0"
+	"go.opentelemetry.io/otel/semconv/v1.36.0/otelconv"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -24,12 +29,34 @@ type logger struct {
 
 	provider             *LoggerProvider
 	instrumentationScope instrumentation.Scope
+
+	selfObservabilityEnabled bool
+	logCreatedMetric         otelconv.SDKLogCreated
 }
 
 func newLogger(p *LoggerProvider, scope instrumentation.Scope) *logger {
-	return &logger{
+	l := &logger{
 		provider:             p,
 		instrumentationScope: scope,
+	}
+	l.initSelfObservability()
+	return l
+}
+
+func (l *logger) initSelfObservability() {
+	if !x.SelfObservability.Enabled() {
+		return
+	}
+
+	l.selfObservabilityEnabled = true
+	mp := otel.GetMeterProvider()
+	m := mp.Meter("go.opentelemetry.io/otel/sdk/log",
+		metric.WithInstrumentationVersion(sdk.Version()),
+		metric.WithSchemaURL(semconv.SchemaURL))
+
+	var err error
+	if l.logCreatedMetric, err = otelconv.NewSDKLogCreated(m); err != nil {
+		otel.Handle(err)
 	}
 }
 
@@ -95,6 +122,9 @@ func (l *logger) newRecord(ctx context.Context, r log.Record) Record {
 		attributeValueLengthLimit: l.provider.attributeValueLengthLimit,
 		attributeCountLimit:       l.provider.attributeCountLimit,
 		allowDupKeys:              l.provider.allowDupKeys,
+	}
+	if l.selfObservabilityEnabled {
+		l.logCreatedMetric.Add(ctx, 1)
 	}
 
 	// This field SHOULD be set once the event is observed by OpenTelemetry.